made app a function component and made a feed page to call the api to show articles
