+++ dashboard/src/pages/FeedPage.js
@@ -1,0 +1,58 @@
+import "../App.css";
+import React, { Component } from "react";
+import ArticleCard from "../components/ArticleCard";
+
+// TODO - have the server filter the results by getting the query in the url in the request.
+const filteData = (data, query) => {
+  if (!query) {
+    return data;
+  }
+
+  return data.filter((article) => {
+    const postName = article.name.toLowerCase();
+    return postName.includes(query);
+  });
+};
+
+class FeedPage extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      articles: [],
+    };
+  }
+
+  componentDidMount() {
+    // GET request using fetch with set headers
+    const headers = { "Content-Type": "application/json" };
+    fetch("https://api.npms.io/v2/search?q=react", {
+      headers: headers,
+    })
+      .then((response) => response.json())
+      .then((data) => this.setState({ articles: data.results }));
+  }
+
+  render() {
+    if (this.props == null) return;
+    const { articles: articles } = this.state;
+    const filteredPosts = filteData(articles, this.props.query);
+
+    return (
+      <div>
+        <dl>
+          {filteredPosts.map((article) => (
+            <dd key={article.package.date}>
+              <ArticleCard
+                title={article.package.name}
+                content={article.package.description}
+              />
+            </dd>
+          ))}
+        </dl>
+      </div>
+    );
+  }
+}
+
+export default FeedPage;
